+++ example/main.cpp
@@ -1,16 +1,10 @@
 #include <hlk/logger/logger.h>
-#include <hlk/logger/commonmessagebuilder.h>
+#include <hlk/logger/basicmessagelayout.h>
 #include <hlk/logger/terminallogginghandler.h>
 #include <hlk/logger/filerotatelogginghandler.h>
-#include <hlk/logger/loggermanager.h>
 #include <hlk/logger/minilogger.h>
 
 int main(int argc, char* argv[]) {
-    // Configure builder
-    auto commonInfoMessageBuilder = std::shared_ptr<Hlk::CommonMessageBuilder>(new Hlk::CommonMessageBuilder());
-    auto commonErrMessageBuilder = std::shared_ptr<Hlk::CommonMessageBuilder>(new Hlk::CommonMessageBuilder());
-    commonErrMessageBuilder->setLoggingLevel(Hlk::CommonMessageBuilder::kError);
-
     // Configure handlers
     auto fileRotateLoggingHandler = std::shared_ptr<Hlk::FileRotateLoggingHandler>(new Hlk::FileRotateLoggingHandler());
     fileRotateLoggingHandler->setLogFilename("log/common.log");
@@ -21,25 +15,22 @@ int main(int argc, char* argv[]) {
 
     // Configure loggers
 
-    Hlk::Logger commonInfoLogger;
-    Hlk::Logger commonErrorLogger;
+    Hlk::Logger<Hlk::BasicMessageLayout> commonInfoLogger;
+    Hlk::Logger<Hlk::BasicMessageLayout> commonErrorLogger;
     
-    commonInfoLogger.setMessageBuilder(commonInfoMessageBuilder);
     commonInfoLogger.pushHandler(terminalLoggingHandler);
     commonInfoLogger.pushHandler(fileRotateLoggingHandler);
 
-    commonErrorLogger.setMessageBuilder(commonErrMessageBuilder);
     commonErrorLogger.pushHandler(terminalLoggingHandler);
     commonErrorLogger.pushHandler(fileRotateLoggingHandler);
 
     // Write logs
-    commonInfoLogger.write("My info message");
-    commonErrorLogger.write("My error message");
+    commonInfoLogger.write("INFO", "My info message");
+    commonErrorLogger.write("ERROR", "My error message");
     
-    auto log = Hlk::LoggerManager::createInstance("my-log");
-    log->setMessageBuilder(commonErrMessageBuilder);
+    auto log = Hlk::Logger<Hlk::BasicMessageLayout>::getInstance("my-log");
     log->pushHandler(terminalLoggingHandler);
-    log->write("My error from registered log");
+    log->write("CUSTOM", "My error from registered log");
 
     Hlk::MiniLogger::write("info", "log/mini.log", "My message from MiniLog");