+++ include/hlk/logger/basicmessagelayout.cpp
@@ -20,28 +20,31 @@
  * 
  *****************************************************************************/
 
-#include "logger.h"
+#include "basicmessagelayout.h"
 
 namespace Hlk {
-
-/******************************************************************************
- * Constructors / Destructors
- *****************************************************************************/
-
-Logger::~Logger() {
-    m_handlers.clear();
-}
 
 /******************************************************************************
  * Methods
  *****************************************************************************/
 
-void Logger::write(const std::string &message) {
-    auto formattedMessage = m_messageBuilder->build(message);
+std::string BasicMessageLayout::formatCurrentTime() {
+    time_t now = time(0);
+    tm *ltm = localtime(&now);
+    char hour[3], min[3], sec[3]; // Two digits and null-terminator
+    sprintf(hour, "%02d", ltm->tm_hour);
+    sprintf(min, "%02d", ltm->tm_min);
+    sprintf(sec, "%02d", ltm->tm_sec);
+    return std::string(hour) + ":" + min + ":" + sec;
+}
 
-    for (size_t i = 0; i < m_handlers.size(); ++i) {
-        m_handlers[i]->write(formattedMessage);
-    }
+std::string BasicMessageLayout::formatCurrentDate() {
+    time_t now = time(0);
+    tm *ltm = localtime(&now);
+    char day[3], month[3]; // Two digits and null-terminator
+    sprintf(day, "%02hhu", ltm->tm_mday);
+    sprintf(month, "%02hhu", ltm->tm_mon + 1);
+    return std::to_string(1900 + ltm->tm_year) + "-" + month + "-" + day;
 }
 
 } // namespace Hlk