+++ include/hlk/logger/loggermanager.cpp
@@ -1,7 +1,0 @@
-#include "loggermanager.h"
-
-namespace Hlk {
-
-std::map<std::string, Hlk::Logger *> LoggerManager::m_loggersByNames;
-
-} // namespace Hlk